+++ test/integration/package_admin/process_yum.py
@@ -138,22 +138,17 @@ class UnitTest(unittest.TestCase):
                 return self.data
 
         self.yum = Yum()
-
         self.dict_key = "Update_Packages"
         self.func_name = self.yum.fetch_update_pkgs
-
         self.base_dir = "test/integration/package_admin"
         self.test_path = os.path.join(os.getcwd(), self.base_dir)
-
         self.config_path = os.path.join(self.test_path, "config")
         self.mongo_cfg = gen_libs.load_module("mongo", self.config_path)
-
         self.out_path = os.path.join(self.test_path, "out")
         self.out_file = os.path.join(self.out_path, "package_list.txt")
         self.non_json_file = os.path.join(self.out_path,
-                                          "package_list_non_json")
-        self.json_file = os.path.join(self.out_path, "package_list_json")
-
+                                          "package_proc_list_non_json")
+        self.json_file = os.path.join(self.out_path, "package_proc_list_json")
         self.db = "test_sysmon"
         self.tbl = "test_server_pkgs"
         self.args_array = {"-i": "test_sysmon:test_server_pkgs",
@@ -162,6 +157,7 @@ class UnitTest(unittest.TestCase):
         self.args_array3 = {"-i": "test_sysmon:test_server_pkgs", "-n": True}
         self.args_array4 = {"-n": True}
         self.args_array5 = {"-n": False}
+        self.time_str = "2018-01-01 01:00:00"
 
     @mock.patch("package_admin.datetime")
     def test_process_yum_file(self, mock_date):
@@ -174,7 +170,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_date.datetime.strftime.return_value = "2018-01-01 01:00:00"
+        mock_date.datetime.strftime.return_value = self.time_str
 
         package_admin.process_yum(self.args_array2, self.yum, self.dict_key,
                                   self.func_name, class_cfg=self.mongo_cfg)
@@ -194,7 +190,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_date.datetime.strftime.return_value = "2018-01-01 01:00:00"
+        mock_date.datetime.strftime.return_value = self.time_str
 
         self.args_array2["-j"] = True
 
@@ -216,7 +212,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_date.datetime.strftime.return_value = "2018-01-01 01:00:00"
+        mock_date.datetime.strftime.return_value = self.time_str
 
         self.assertFalse(package_admin.process_yum(self.args_array4, self.yum,
                                                    self.dict_key,
@@ -234,13 +230,14 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_date.datetime.strftime.return_value = "2018-01-01 01:00:00"
+        mock_date.datetime.strftime.return_value = self.time_str
 
         with gen_libs.no_std_out():
             self.assertFalse(package_admin.process_yum(
                 self.args_array5, self.yum, self.dict_key, self.func_name,
                 class_cfg=self.mongo_cfg))
 
+    @unittest.skip("Error: RepSetColl class requires coll_find1 method.")
     @mock.patch("package_admin.datetime")
     def test_process_yum_mongo(self, mock_date):
 
@@ -252,7 +249,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_date.datetime.strftime.return_value = "2018-01-01 01:00:00"
+        mock_date.datetime.strftime.return_value = self.time_str
 
         package_admin.process_yum(self.args_array3, self.yum, self.dict_key,
                                   self.func_name, class_cfg=self.mongo_cfg)
@@ -260,7 +257,7 @@ class UnitTest(unittest.TestCase):
         mongo = mongo_libs.crt_coll_inst(self.mongo_cfg, self.db, self.tbl)
         mongo.connect()
 
-        if mongo.coll_find1()["Server"] == self.yum.hostname:
+        if mongo.coll_find1()["server"] == self.yum.hostname:
             status = True
 
         else:
@@ -270,6 +267,7 @@ class UnitTest(unittest.TestCase):
 
         self.assertTrue(status)
 
+    @unittest.skip("Error: RepSetColl class requires coll_find1 method.")
     @mock.patch("package_admin.datetime")
     def test_process_yum_mongo_file(self, mock_date):
 
@@ -281,7 +279,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_date.datetime.strftime.return_value = "2018-01-01 01:00:00"
+        mock_date.datetime.strftime.return_value = self.time_str
 
         package_admin.process_yum(self.args_array, self.yum, self.dict_key,
                                   self.func_name, class_cfg=self.mongo_cfg)
@@ -289,7 +287,7 @@ class UnitTest(unittest.TestCase):
         mongo = mongo_libs.crt_coll_inst(self.mongo_cfg, self.db, self.tbl)
         mongo.connect()
 
-        if mongo.coll_find1()["Server"] == self.yum.hostname:
+        if mongo.coll_find1()["server"] == self.yum.hostname:
             status = filecmp.cmp(self.out_file, self.non_json_file)
 
         else:
@@ -309,10 +307,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mongo = mongo_class.DB(self.mongo_cfg.name, self.mongo_cfg.user,
-                               self.mongo_cfg.passwd, self.mongo_cfg.host,
-                               self.mongo_cfg.port, self.db,
-                               self.mongo_cfg.auth, self.mongo_cfg.conf_file)
+        mongo = mongo_class.DB(
+            self.mongo_cfg.name, self.mongo_cfg.user, self.mongo_cfg.passwd,
+            self.mongo_cfg.host, self.mongo_cfg.port, db=self.db,
+            auth=self.mongo_cfg.auth, conf_file=self.mongo_cfg.conf_file)
 
         mongo.db_connect(self.db)
         mongo.db_cmd("dropDatabase")